restructure push buttons handler
